Include the current time of the panic in the panic output
Show both the real and the monotonic time in the panic output when the
panic is created. This is useful to have something to compare other
panic timestamps against.

The existing timestamp in the panic output is recorded by the
management process at the time of receiving the panic from the child
process. Due to processing and possible core dump creation, this time
isn't very useful for comparison purposes.