Overhaul statistics aggregation: Send worker stats to a separate thread
Previously, statistics aggregation was a three-level process:

1) workers collected per-thread statistics

2) When wthread_stats_rate was reached or a worker went idle,
   statistics were aggregated on the pool level

3) Only when a worker was idle were the pool level statistics added to
   the global aggregation

This could lead to long delays between the time some statistic was
generated and it being reported, in particular under full load. Global
statistics were never updated as long as all worker threads were busy
at all times.

We now simplify this into a two-level process: As before, workers keep
their own statistics, but when it is time to aggregate, we send them
off to a separate thread, which periodically adds all received stats
to the global counters.

The parameter thread_stats_rate is replaced by thread_stats_latency,
which specifies a latency goal for statistics aggregation. The default
of 1s is motivated by the default varnishstat update interval.