Introduce a parameter value for extrachance retries
There seem to be conditions in the extrachance logic under which Varnish
could retry a backend request indefinitely. This commit introduces a new
parameter, gethdr_extrachance, allowing to set an upper bound for the
amount of retries in vbe_dir_gethdrs.

The default value for the newly introduced parameter is 1, setting it to
0 disables the feature entirely.