Collect vtest failed tests logs more reliably
Instead of randomly looking for the log files, look for them where we
expect them to land. For instance, after a git pull, a distclean
operation may not remove the dist directory if Varnish's version
changed because it would reconfigure itself on the fly.

Instead of reading the output of `make distcheck`, look for failing
tests in the dist directory.