Don't inherit a top-level's backend
When a top-level VCL switches to a label, the `req` variable is
inherited as is, including `req.backend_hint`. This happens even
if it wasn't set explicitly.

Since Varnish 4.1, backends are specifically owned by VCLs instead
of being shared as they were before. During a VCL selection, the
director hint is now set to the label's default.

There is also a test case proving this comment wrong from
`vcl-separate.rst`:

> We have to have a backend, even if we do not use it