Autocrap de-cargo-culting
We don't need AC_PROG_CPP, because we don't use the C preprocessor
standalone for anything.

It's not obvious that anybody needs AC_PROG_MAKE_SET ever, it looks
like a band-aid for some crappy Win98-related tinker-toy-compiler.

Amazing (or not...) that nobody in the autocrap community ever
got the idea to add "What made AC_FOOBAR necessary" explanations
to their documentation or changelogs.