Dedicated autoconf macros for libraries detection
They live as private macros in varnish.m4, which is now included in the
Varnish build during the autogen phase. This way it is now possible to
dogfood on macros we distribute downstream. Considering how simple they
are, they could be made public macros for the convenience of VMOD or VUT
writers.

We don't need to have the LIBS-saving-dance in configure.ac for every
single dependency.