fix enum check in vcc: do not accept values valid for other arguments
VCC was silently accepting enum values valid for other arguments
following in the argument list as well as other identifiers in the
argument spec (for example "ENUM"). Consequently, wrong enum values in
VCL were not detected at VCC time and passed to vmod functions, which,
in the best case, would detect the error (and, if following the
varnish good practice, panic on a failed assertion).

This is another forgotten case since the enum list was changed to be
terminated by \1 in a78efad8002895e6097aeb6b1daeac0f6108b9a9: vcc_expr
would just loop over the \1 terminator up to the final \0 terminator
at the end of the argument spec.