run all the rst processing in builddir to fix out-of-tree builds
I am not entirely happy about this and we should try to find a better solution
to this chicken-egg-ish situation:

- part of our documentation is hand-written (in git/srcdir), part is auto-
  generated

- because we add generated rsts to the dist, they are not in srcdir when
  building from git, but they are when building from a dist (tarball)

- sphinx wants all files in one doctree - there seems to be no include path

I tend to lean towards removing generated rsts from the dist to unify the
src tree for git vs. dist builds, but I'd like to avoid too much change at
this point due to other priorities.