distcheck: use the original CFLAGS for the inner configure
In particular with --enable-developer-warnings, we select strict compiler
options for CFLAGS, which are too strict for the configure built-in conftest.c
code.

Make distcheck calls configure again for an out-of-tree build, so it would
see the CFLAGS the "outer" configure built. To avoid this situation, we
restore the original CFLAGS for the inner configure.

I wonder if there is a better, more autoconf-ish, solution to this problem.