Add missing entries in the Huffman decoding tables
The very last octet, byte3_pref1111, although coded on 6 bits actually
needs the 8 of them. Because if we get past the legitimate 30 bits of
the worst codes, we find ourselves with spurious EOS entries. We can
distinguish EOS entries by giving them a length of 0 bits.

Why do we need 32 bits you ask? Because the lookup tables contain
partial codes aligned left. So the biggest possible code 0x3ffffffe
would be represented as 0xfffffff8 instead.

The last octet 0xf8 tells us that we need a lot more than 64 entries,
the rest is amusing bit twiddling hacks to determine how many times
each entry is repeated in the table. That is to say, a mask.