Fix ESI Parsing when VEP_Parse is called on tag which shares a prefix with another
Example for the case addressed here is an html comment which is not esi.

	1st call on !--e
	2nd call on !--end

The first call could match !--esi:, so vep_match() does not return with
VEP_NOTMYTAG, but with null, so characters up until 'e' are marked.

For the second call, we need to consider the case that characters after
the end of the tag could already be marked.

Fixes #1988