Introduce a `vtest-clean` make target
When a `distcheck` fails, the $(distdir) is not removed and this is how
we can collect test results. This directory is also left untouched by
the `distclean` target run from `autogen.des`.

So if the next commit checked by vtest fails before running inside the
$(distdir), it will report failures from the previous run. To avoid that
we can use $(am__remove_distdir) that deals with permissions but may not
be stable API.