Introduce macros to bulk-check CFLAGS
As usual, reusable macros are put in `varnish.m4` which is shipped
downstream. Although CFLAGS checks rely on `AX_CHECK_COMPILE_FLAG`
it is safe to do so as long as VMOD or VUT projects don't use the
new macros: they are documented as private anyway.

The idea is to leave it to compiler checks to decide whether a flag
can be used, regardless of the overall target. It will be decided at
configure time instead. It would be nice if we could trust compilers
but the normal course of action seems to be quirky.

Some SunCC-specific checks are still around for now but ideally they
will go away too.