Rework the -f/-b logic so we can take multiple -f arguments.
If you give multiple -f arguments, the resulting VCL's will be
"boot0", "boot1", ... "boot668", "boot"

And the last one will be the active one.

The plan is to also allow both specification of an (optional) label and
which VCL is to be the active one, but I want to think about if we should
shoehorn that into -f or use a different option letter.