add std.send_100_continue to optimize request body latency
Postponing the 100 Continue response until when we start working
the request body is correct, but this may not happen until we have
opened a backend connection or issued other io.

This vmod function allows VCL authors to optimize request body
processing by signalling the 100 Continue to the client as soon
as a decision has been made in VCL that a request body is to be
received. Doing so may reduce latency by motivating the client
to fill its send window before we actually start reading the body.