Add a setenv command to varnishtest
This lets you set environment variables in a test case.

Usage is

    setenv "FOO" "bar"

which will set FOO=bar if FOO does not already exist.

There is also an optional '-force' argument, which will overwrite the
current variable if already set:

    setenv -force "FOO" "baz"

The force argument corresponds to passing a non-zero third argument to
setenv(3).