Add a setenv command to varnishtest
This lets you set or change environment variables in a test case.

Usage is

    setenv FOO "bar baz"

The above will set the environment variable $FOO to the value
provided. There is also an ``-ifunset`` argument which will only set the
value if the the environment variable does not already exist:

    setenv -ifunset FOO quux

The ifunset argument corresponds to passing zero as the third argument
to setenv(3).