Get accurate byte counters
There was a regression from Varnish 4.0 to 4.1, where the response
bytes was counted as the number of bytes fed to the outgoing write
vector, rather than the bytes that was actually handed off to the OS'
socket buffer. This would cause for many cases the complete object
size counted as transmitted bytes, even though the client hung up the
connection early.

To fix this, two new fields has been added to the ReqAcct and
BereqAcct log lines. These fields show the number of protocol bytes
that were read or written to the socket. These bytes will include any
extra protocol overhead (e.g. chunked encoding).

varnishncsa has been changed to use these new counters in its byte
count output. For the %b formatter (body bytes sent), the output will
be the smallest of the response body bytes, or the protocol response
bytes. This is to give the expected output for known object sizes
without inflation on connection hangups.