Back to Expect handling according to rfc7231
Revert a regressive nit from bdd84ffc5300c25ca8c13b3aefa71e05cad99d26 which
has nothing to do with H/2

https://tools.ietf.org/html/rfc7231#section-5.1.1

   Requirements for servers:

   o  A server that receives a 100-continue expectation in an HTTP/1.0
      request MUST ignore that expectation.

So if Expect exists but is not "100-continue", we send the 417 irrespective
of the protocol version.

But if is is 100-continue and the protocol is HTTP/1.0 or below, we ignore
the header.