Roll back the fix for #1206, so that client delivery solely decides whether to set client response 304, regardless of the pass state of the backend fetch. That is, set resp.status 304 (at ESI level 0) iff req.http.{INM,IMS} matches resp.http.{ETag,Last-Modified}.
This means that the client will get a 304 response in the pass
cases where the backend sets ETag or Last-Modified to match INM/IMS,
but nevertheless declines to set beresp.status to 304. This behavior
is consistent for the three kinds of pass: "ordinary" pass (return(pass)
in vcl_recv), beresp.uncacheable ("hit-for-miss"), and the recently
revived hit-for-pass (return(pass(DURATION)) in vcl_backend_response).

r02042.vtc shows how to restore the behavior from #1206, by unsetting
INM/IMS in vcl_pass (for ordinary pass) or vcl_miss (for hit-for-miss).