Don't protect the barriers list with a lock
A command like `barrier r1` would fail while holding the lock because of
the invalid name, but the reset of the test case would then deadlock. It
leads to a timeout like we see sometimes on VTEST so getting rid of them
may just require to review when `vtc_fatal` is called while locking.

In this very case we can only create barriers in the top thread so the
lock is useless.