Fail name checks early in varnishtest
It closes a race for barriers because the list could be updated by the
top thread while walked by another thread. It also saves useless lookups
by rejecting names earlier.

This is not the case for servers because a) they have proper locking of
the servers list and b) there is the dispatch special case. It would
otherwise be harmless to do it for the server too because dispatched
servers don't depend on user input.