Fix rendering of varnishd -? in sphinx
Apparently rst2man and sphinx don't have the same opinion on what an
option is or isn't. Looking at the RST spec it doesn't seem to allow a
question mark [1] in an option list. Among the various kind of options
we can see:

    Short POSIX options consist of one dash and an option letter.

To work around that I introduced a crude hack using sed. It satisfies
my OCD but I'm not even remotely proud of that.

[1] http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#option-lists