Do not update mgt's mirrored VCL temp state on child temp state failures
The child expects mgt to be consistent in its VCL state changes,
e.g. not ask the child to 'use' a VCL that isn't warm. Not doing so
will cause the child to assert (see asserts in
cache_vcl.c:ccf_config_use). This could trigger because the mgt
erranously would set its recorded temp regardless of the child's reply
in mgt_vcl_setstate.

This patch changes it to only record the new temp in mgt if the child
replied success. If the child isn't running the new temp is always
recorded.

Slightly update v00044.vtc test case to reflect the need to update the
state before restarting the child.

Ref: https://github.com/varnishcache/varnish-cache/pull/2273