Support multi-lib systems for out-of-tree builds
On systems where both 32 and 64-bit libraries can co-exist, simply
inheriting Varnish's prefix is not enough because VMODs might land
in the wrong directory. For instance, on rhel you will find both
/usr/lib and /usr/lib64.

This is usually not a problem with package managers, because they
usually take care of everything at configure time, for example RPMs
may use a %configure macro that defines all well-known directories.
It can however be overlooked when a VMOD is manually built from
source.