Fix an issue with hpack table shifting
When inserting a new hpack table entry using a referenced name, and
the referenced name would be removed from the table as a result of the
insertion, there was a bug in the memory move/copy operations that
would memmove a too large area. This wouldn't cause problems because
we never used the full 32 bytes that the standard sets as per element
overhead.

Fix it, and the same time simplify the logic to be easier to read. Add
a test case for trimmed indexed name insertion where the table isn't
empty after removing the indexed entry.