Polish verbose compliance in varnish.m4
This is a cosmetic tweak, it doesn't change any behavior. No need to
bump the serial for that. The benefits are a better output of silent
make executions, and the preservation of configured variables like
$(PYTHON) all the way downstream.

A word on this stunt:

    \$(A""M_V_GEN)

The backslash is here to escape the dollar sign, nothing too surprising.
However the expression between the brackets is rejected by autoconf
because it looks like an undefined macro. To prevent that, the string is
closed and reopen at once to break the token-inside-brackets construct.