Prevent storage backends name collisions
Because of Transient storage being a special-case stevedore, the
STV_Foreach logic would only work after the child's startup once
all storage backends are defined. So during the setup, collisions
would only be detected for the Transient storage.

The underlying STV__iter function is now pretty much like a plain
VTAILQ_FOREACH loop with an additional CHECK_OBJ_ORNULL step on all
iterations.

As a result, the transient storage is appended to the list at the
end of the manager's setup and is now reported last in the CLI's
storage.list command.

Error messages related to the -s option are slightly more helpful.

Fixes #2321