Remember to actually record the length of variable length attributes
A variable length attribute (OA_HEADERS and OA_VARY) are allowed to be
overwritten if the new data has the exact same length as the previous
data. To check this, its length is recorded. But the length wasn't
recorded, and the code would treat it as a new attribute.

The reason this hasn't caused issues is that the code looking at
OA_HEADERS and OA_VARY ignores the stored attribute length, and that
these aren't ever overwritten.