sync idle_send_timeout and send_timeout documentation with reality
The name idle_send_timeout does not seem to describe well where the timeout
applies (individual write()/writev() calls for arbitraty amounts of data).

The fact that the explanation is complicated might be an indication that a
different timeout implementation be more appropriate.

HTTP2 should probably implement send_timeout somewhen.