First babystep of the new selfdocumenting VSC counters.
The .VSC file steals .RST's directive syntax, but it is not a RST
file, and it is not processed as such.  Maybe later it will become
a .RST file, but that is TBD.

A C-struct for use in varnishd is built, (along with a JSON structure
which describes the counters.)

For now that structure is identical with what comes out of
include/tbl/vsc_f_main, so nothing special happens in the VSM yet.