Prevent storage backends name collisions
This is a backport of 986763198 by Dridi Boukelmoune, but this patch
is a bit different from the original.

Now name collisions with transient will be reported in a sane way
instead of producing an assert. Normal name collisions are also more
readable.

The test case is unmodified, thanks to resent back porting efforts in
varnishtest.

Fixes #2321

Conflicts:
	bin/varnishd/storage/mgt_stevedore.c