Introduce the vmod vtc for varnishtest
This is moving some of the generally useful functions of vmod debug to a
new one that is installed for use out of the Varnish source tree. Things
weren't just moved from one source file to another, some functions were
sligthly renamed, and the behavior of `barrier_sync` was changed to
adhere to the "universal fail" and no longer returns a boolean.

While at it, I added missing __match_proto__ macros to vmod debug.

Closes #2276