Update some uses of VTCP_*name when the socket is UDS.
* The ADDR buffer length is now 128, more than enough for a UDS path.
* VTCP_*name, if called for a UDS socket, puts the path in the addr
  buffer and "<none>" in the port buffer.

Note however that getsockname and getpeername may return unnamed
sockaddr_un's (garbage in the sun_path field), depending on whether
the socket is bound or connected.

UDS bound -> getpeername returns unnamed sockaddr_un
UDS connected -> getsockname returns unnamed sockaddr_un

For these reasons, we try to avoid VTCP_*name for UDSen when it's
possible to obtain the VSA, and call VSA_Path() instead.

This is apparently not possible where BackendStart is emitted to
the log, since it is called in a function with structs that don't
seem to have a way to navigate back to struct backend, where the
UDS VSA lives. So it calls VTCP_hisname() (correct for a connected
socket to a backend), and "/path <none>" is emitted for BackendStart.

Uses of VTCP_*name not covered in this commit:
* PROXY protocol
* mgt/CLI
* varnishtest