Expose to VCL whether or not a fetch is a background fetch (bgfetch)
The use case are cluster requests: Intra-cluster bgfetches should
trigger a synchronous fetch on the peer-varnish in order to avoid
additional short-lived / expired objects being created. Or, in
other words, a bgfetch should actually get a fresh object and not
one in grace from another cache.

Merges #2376

This is a back port of d7f8b531bd63fcb5b.

Conflicts:
	bin/varnishd/cache/cache_fetch.c
	doc/changes.rst
	include/tbl/bo_flags.h