New req_dropped counter similar to sess_dropped
This one is for h2 streams, so the thread pools needed to learn about
"stream" tasks. To properly test this counter, thread_pool_timeout is
now disregarded when the vtc_mode debug flag is set. A minimum value
of 10s is unrealistic for the test suite, it is now hard-coded to 0.5s
when ran through varnishtest.