Ignore req.ttl when keeping track of epired objects
The goal of `req.ttl` is to allow lower requirements for known
transactions (since this has to be done in VCL) but objects with
an age higher than `req.ttl` would always be considered expired
during lookup even if their actual TTL (obj.ttl) is still positive.

Ignoring `req.ttl` gives a better control over explicit refreshes
made to optimize latency and greatly reduces the risk of running
into #1799 (for a subset of use cases).

Closes #2422