Move the cli buffer to VSB (from stack)
If the cli buffer is allocated on the stack, the mgt process will
crash hard for big values of this parameter. Instead we allocate on
the heap.

The new cli buffer an "auto" VSB. This means that the varnish
parameter cli_buffer is ignored from now on, and the manual marks it
as deprecated.

Closes: #2382