Enforce sequence requirement of header block frames
This commit enforces that a HEADERS frame without an END_HEADERS flag
must be followed by a CONTINUATION frame on the same stream.

Sending interleaved/out of sequence header or continuation frames will
now be treated as a PROTOCOL_ERROR, as per rfc7540.

This also fixes the panics seen in #2387.

Fixes: #2387