Pass delivery abandoned does not qualify as an error
... so log it under the Debug tag.

FetchErrors should be actual errors which can be addressed. In this case,
nothing is wrong in any way, the fact that we abort a fetch if we don't
need the body is a varnish internal optimization (which makes sense, but
comes at the cost of closing a connection).

Merges #2450

Conflicts:
	bin/varnishd/cache/cache_fetch.c