Replace FetchError "no backend connection" with something more useful
Previously, a FetchError in vbe_dir_getfd() was too general to be
useful. Now we log an error with the same identifier as the
corresponding statistic:

* unhealthy
* busy (max_connections reached)
* fail (no connection for real)

For these, we also log the backend name because we have not yet issued
a BackendOpen VSL.

A failed WS_Alloc is logged as "out of workspace".