Make vrt.h include <stddef.h> and <stdint.h>
Originally we decided that the output of VCC should be 100%
stand-alone and therefore contain no #includes at all.  This
was hoped to avoid unspecified trouble with C-compilers at runtime.

But C99 is old enough to drink now, so we move forward.

The script in tools/include_wash.py checks *.c files and
complains about violations of our intended #include orders.