Add cache_hit_grace counter
The cache__hit_grace counts the number of grace hits. To be precise,
it counts the number of times lookup returns an expired object,
but vcl_hit is called and decides to return(deliver).

Every time cache_hit_grace is incremented, cache_hit is also
incremented (so this does not change the cache_hit counter).