take v1l memory from the thread workspace (again)
... as it was the case before 69d45413bb1fca69cb487b0098f1a7323114fdf7
and as documented.

The motivation is to remove the reservation from req->ws during
delivery, but actually line delivery memory should not come from the
request space - as originally designed:

- We avoid requiring to have an obscure surplus of workspace_client
  for delivery

  - which is also added for every subrequest though not required there

- We get predictable performance as the number of IO-vectors available
  is now only a function of workspace_thread rather than the amount of
  memory which happens to be available on the request workspace.

As a sensible side effect, we now also fail with an internal 500 for
workspace_session and workspace_thread overflows in addition to the
existing check on workspace_client for completeness.

Due to the dual-use of the thread workspace (delivery + scratch space
in vrt and elsehwere), we need to temorarily switch the thread
workspace to a new instance when it is reserved (during ESI processing).