Add cache_hit_grace counter
The counter cache_hit_grace counts the number of grace hits. To be
precise, it counts the number of times lookup returns an expired
object, but vcl_hit is called and decides to return(deliver).

Every time cache_hit_grace is incremented, cache_hit is also
incremented (so this commit does not change the cache_hit counter).

This is a back port of 1d62f5da43a245966faff82f3ee3e6cff46db0f2.