VIP12 vmod function/method polymorphism
Ref: https://github.com/varnishcache/varnish-cache/wiki/VIP12:-vmod-polymorphism-(for-type-conversions)

We now allow definition of vmod function/methods by the same name but
different signatures.

VCC file:

For $Function/$Method names, we add a "poly suffix" {C_SUFFIX} which
is appended to the function names for the C-Interface, but not for the
VCL interface.

vmodtool:

The change to extract the poly suffix is straight forward. The
RST generation is unchanged for now. All information is contained,
but we might want to improve the documentation format for polymorphic
functions later.

vcc:

We hang polymorphic variants of the same symbol name as a chain of
children under the first symbol found. While using just one layer
below a primary symbol would be a bit more efficient for symbol
insertion, this way we do not need to break the rule of symbol names
being unique at each level.

In vcc_Eval_SymFunc(), the first matching variant is now looked up by
successively trying vcc_func() until success or symbols exhausted.

This implementation differs from the VIP in some regards:

* "for polymorphic arguments, STRING always needs to be supported"

I cannot see any more why we would want to be that strict. For string
arguments, a polymorphic function (as any function/method) effectively
takes any argument type through string folding, and a vmod might want
to not allow that.

* "at vcl compile time, vcc finds the best match ** arbitration
   left->right best match (e.g. stringification of rightmost arguments
   preferred)"

Such heuristic appears pretty arbitrary. Instead, the current
implementation just uses the first match found, in the order of
definition in the vcc file. This should give vmod authors maximum
flexibility - including varying argument counts.