Trade space for time in setting up VSC segments.
Originally it was the plan that the JSON would be gzip'ed in the
VSM, but to avoid exposing VGZ in the API we didn't, but kept the
gzip'ed JSON in the generated VSC code because why not?

Well now we know why not:  It wastes a lot of time for very little space,
in particular if you have many backends.

Provides some relief for: #2513 & #2518