add return(restart) from vcl_recv
This is already possible with some overhead (for example with a
de-tour via vcl_pass), but I see absolutely no reason why we should
not allow direct restarts from recv.

The use case is to rewrite the request and re-start processing of the
rewritten request as if it was received as such. This is also possible
already with some vcl sub juggling, but is complicated by the VCL call
loop detection. Use of restarts avoids this complication and is safe
due to max_restarts.