Use the request's timestamp t_req for ttl calculations on the object
To ensure that the calculation of the ttl is equal in vcl and the C
code, the vcl "variable" for an object's ttl is now calculated from
the request timestamp, not the last timestamp recorded. For the
backend side, we continue to use an as recent as possible timestamp
for the ttl variable.

The test case is called r02519.vtc since the problem only became
visible after the fix initially proposed in Pull Request #2519.

The change to c00012.vtc is due to the fact that we expect o_ttl to
be exactly 0.000 in normal cases.